Enable use of subset fonts
All the included fonts were causing the repo size and the dist
size to increase substantially. Not all glyphs included in the
.ttf files will actually be used in real-life. As such, this
commit refactors the fonts to subset fonts with limited glyphs.
NotoSans-Regular has been left as-is to support all unicode
characters as a fallback font.

Note that there is also an issue with FPDF2, that causes fallback
fonts to not work sometimes if they have a space in the name. As
such, I have changed all font names to remove spaces.
See https://github.com/py-pdf/fpdf2/issues/1535 for more details.